+++ NackEngine/math/ProbDensity.cs
@@ -1,13 +1,10 @@
 ﻿using NackEngine.core.space;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace NackEngine.math
 {
     public interface ProbDensity
     {
-        double Value(NVector direction);
+        float Value(NVector direction);
         NVector Generate();
     }
 }
